Requisito 9

+++ models/salesModel.js
@@ -3,7 +3,13 @@ const connection = require('./connection');
 
 const registerSale = (sale) =>
   connection()
-  .then((db) => db.collection('sales').insertOne({ itensSold: sale }))
+  .then((db) => {
+    sale.forEach((item) => {
+      db.collection('products')
+      .updateOne({ _id: ObjectId(item.productId) }, { $inc: { quantity: (-1) * item.quantity } });
+    });
+    return db.collection('sales').insertOne({ itensSold: sale });
+  })
   .then((result) => result.ops[0]);
 
 const getAll = () =>
@@ -23,10 +29,22 @@ const updateSale = (id, itensSold) => {
   return { _id: id, itensSold };
 };
 
-const deleteSale = (id) =>
-  connection()
+const deleteSale = (id) => {
+  if (!ObjectId.isValid(id)) return null;
+  const results = [];
+  return connection()
   .then((db) => db.collection('sales')
-  .deleteOne({ _id: ObjectId(id) }));
+  .findOne(ObjectId(id))
+  .then((data) => {
+    if (data === null) return null;
+    data.itensSold.forEach((item, index) => {
+      results[index] = db.collection('products')
+      .updateOne({ _id: ObjectId(item.productId) }, { $inc: { quantity: item.quantity } });
+    });
+    return Promise.all(results).then(() => db.collection('sales')
+    .deleteOne({ _id: ObjectId(id) }));
+  }));
+};
 
 module.exports = {
   registerSale,
